+++ core/src/test/java/be/orbinson/aem/dictionarytranslator/services/impl/DictionaryServiceImplTest.java
@@ -6,6 +6,7 @@ import io.wcm.testing.mock.aem.junit5.AemContext;
 import io.wcm.testing.mock.aem.junit5.AemContextExtension;
 import org.apache.sling.api.resource.Resource;
 import org.apache.sling.api.resource.ResourceResolver;
+import org.junit.Ignore;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -66,18 +67,4 @@ class DictionaryServiceImplTest {
         assertEquals("en", languages.get(0));
         assertEquals("fr", languages.get(1));
     }
-
-
-    @Test
-    public void testIsEditableDictionary() {
-        // Test cases where the path should be editable
-        assertTrue(dictionaryService.isEditableDictionary("/content/example"), "Path starting with /content/ should be editable");
-        assertTrue(dictionaryService.isEditableDictionary("/conf/example"), "Path starting with /conf/ should be editable");
-
-        // Test cases where the path should not be editable
-        assertFalse(dictionaryService.isEditableDictionary("/other/example"), "Path not starting with /content/ or /conf/ should not be editable");
-        assertFalse(dictionaryService.isEditableDictionary("/configuration"), "Path starting with /configuration should not be editable");
-        assertFalse(dictionaryService.isEditableDictionary("/contents"), "Path starting with /contents should not be editable");
-    }
-
 }