+++ CHANGELOG.md
@@ -10,10 +10,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - AEM 6.5 support [#62](https://github.com/orbinson/aem-dictionary-translator/issues/62)
+- Added prerequisites to README.md [#65](https://github.com/orbinson/aem-dictionary-translator/issues/65)
 
 ### Fixed
 
 - Errors on compilation of dictionary clientlib [#61](https://github.com/orbinson/aem-dictionary-translator/issues/61)
+- Action bar for read only dictionaries shows only valid actions [#42](https://github.com/orbinson/aem-dictionary-translator/issues/42)
+
+### Changed
+
+- Only single dictionary can be selected [#62](https://github.com/orbinson/aem-dictionary-translator/issues/62)
 
 ## [1.2.1] - 2024-11-26